requirements.txt
# Production
fastapi
alembic
sqlmodel
pandas
psycopg2-binary
uvicorn
asyncpg
greenlet
python-dotenv
redis
requests
# Development
pytest
pytest-anyio
pytest-asyncio
pytest-cov
httpx
black
pylint
mdformat
isort
autoflake
sqlalchemy-utils
asgi_lifespan